+++ changelog.md
@@ -18,6 +18,7 @@ Don't publish to MQTT after every change - add small debounce to it
 - Added `/mqtt_config` endpoint to configure mqtt.
 - Added navigation to the html pages and made use common headers
 - Added `current_ms` counter to status page
+- Add decoding for light configured messages 
 
 ### Changed
 - Pulled out led functions to helper file.
@@ -30,7 +31,7 @@ Don't publish to MQTT after every change - add small debounce to it
 - Log out the IP address correctly when it has one.
 - Improved the display of Spa and Pool temps in home assistant box instead of slider
 - Fixed the display of pool light in home assistant
-- Only show channels that are actually configured.
+- Only send discovery messages for channels and lights that actually are in use.
 
 
 ### Removed